YTU-CE Semantic Web Graduate Course- HW Assignment 1- XSD, XML
What is this project about?
In this project, I have tried to show the structure of CBRT in XML
format according to a certain XSD.
What are the files about?
Instructions of the professor (Dr. Mehmet Aktaş) could be found in
instuctions.pdf
Tcmb.xsd contains xml schemas, I want to put some xml files which use
this schema too.
The JAVA project verifies and parses the xml files both with DOM and SAX
parser and queries xml with XPATH.
There is also XSLT a program to convert xmls to html.
I put a project report that includes my experiences and a brief
comparison of Dom and Sax parsers effectiveness.
Enjoy!